Split off folderTools
